* update EasyAnimate train/infer executors to an env-manager-based form and erase the vbench from the sandbox deps
