+++ configs/demo/bench/vbench_eval.yaml
@@ -1,5 +1,9 @@
 type: vbench_video_evaluator
 
+# env related
+env_name: 'vbench'
+env_manager: 'conda'
+
 # The vbench prompts for video generation. Use VBench_full_info.json for full eval.
 full_json_dir: ./tools/mm_eval/vbench_metrics/VBench_mini_info.json
