+++ pyproject.toml
@@ -148,18 +148,6 @@ sandbox = [
     "fire",
     "pyspark",
     "modelscope[framework,nlp]",
-    "deepspeed",
-    "safetensors",
-    "timm",
-    "xformers",
-    "decord",
-    "opencv-python",
-    "omegaconf",
-    "imageio[ffmpeg,pyav]",
-    "tensorboard",
-    "diffusers==0.27.0",
-    "transformers>=4.37.2",
-    "func_timeout",
 ]
